+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/domain/repository/ReportedDemandAndCapacityNotificationRepository.java
@@ -20,10 +20,14 @@ SPDX-License-Identifier: Apache-2.0
 
 package org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.repository;
 
+import java.util.Collection;
+import java.util.List;
+import java.util.UUID;
+
 import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model.ReportedDemandAndCapacityNotification;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ReportedDemandAndCapacityNotificationRepository extends DemandAndCapacityNotificationRepository<ReportedDemandAndCapacityNotification>{
-
+    List<ReportedDemandAndCapacityNotification> findByNotificationIdIn(Collection<UUID> notificationIds);
 }